+++ apps/openmw/mwmechanics/repair.cpp
@@ -44,12 +44,12 @@ void Repair::repair(const MWWorld::Ptr &itemToRepair)
 
     float toolQuality = ref->mBase->mData.mQuality;
 
-    float x = (0.1 * pcStrength + 0.1 * pcLuck + armorerSkill) * fatigueTerm;
+    float x = (0.1f * pcStrength + 0.1f * pcLuck + armorerSkill) * fatigueTerm;
 
-    int roll = static_cast<float> (std::rand()) / RAND_MAX * 100;
+    int roll = static_cast<int>(static_cast<float> (std::rand()) / RAND_MAX * 100);
     if (roll <= x)
     {
-        int y = fRepairAmountMult * toolQuality * roll;
+        int y = static_cast<int>(fRepairAmountMult * toolQuality * roll);
         y = std::max(1, y);
 
         // repair by 'y' points
